Remove duplicate code for RoundLayoutValue
